requirements.txt
streamlit
transformers
torch
spacy
scikit-learn
pandas
plotly
wordcloud
matplotlib
Pillow
